requirements.txt
# Core ML/DL
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.2.0
tensorboard>=2.13.0
torchaudio>=2.1.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

#enviromemt var
python-dotenv

eventlet


# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
fastapi>=0.110.0
flask>=2.3.0
Flask-CORS==4.0.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional for wildlife detection
albumentations>=1.3.0  # Advanced augmentations
imutils>=0.5.4         # Image utilities
pycocotools>=2.0.6     # COCO evaluation

gunicorn==21.2.0
python-dotenv==1.0.0
psycopg[binary]
psycopg_pool
Werkzeug==2.3.0

#Authentication
bcrypt=5.0.0 
pyjwt=2.12.1